# evaclub-ui/src/app/services/models/event-response.ts
/* tslint:disable */
/* eslint-disable */
export interface EventResponse {
  defaultLayout?: boolean;
  enabled?: boolean;
  eventCover?: string;
  eventDate?: string;
  eventId?: number;
  name?: string;
}